+++ CHANGELOG.md
@@ -60,6 +60,7 @@
 - [9021](https://github.com/vegaprotocol/vega/issues/9021) - Add rejection reason to stop orders.
 - [9012](https://github.com/vegaprotocol/vega/issues/9012) - Add governance to update the referral program.
 - [9077](https://github.com/vegaprotocol/vega/issues/9077) - Ensure liquidity `SLA` parameters are exposed on markets and proposals
+- [9046](https://github.com/vegaprotocol/vega/issues/9046) - Send event on referral engine state change.
 
 ### 🐛 Fixes